pyproject.toml
[project]
name = "neuro_pilot"
version = "1.0.0"
description = "NeuroPilot: Unified E2E Autonomous Driving Framework (Multitask, World-Model Ready)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "albumentations>=1.3.0",
    "tqdm",
    "numpy",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "opencv-python",
    "pyyaml",
    "onnx",
    "onnxscript>=0.5.7",
    "pydantic>=2.0",
    "rich",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.21",
    "torchmetrics>=1.8.2",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "pycocotools>=2.0.11",
    "loguru>=0.7.2",
    "omegaconf>=2.3.0",
    "onnxruntime<1.24.0",
    "psutil",
]

[tool.setuptools]
packages = ["neuro_pilot"]

[project.scripts]
neuropilot = "neuro_pilot.entrypoint:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy"
]
